Fix broken modals/dropdowns, rebrand to EchoCrew, and repair PHP8/composer setup
UI/branding:
- Rebrand login page and app-wide footer from vtiger to EchoCrew;
  remove the live call to vtiger's marketplace API that was pulling
  vtiger marketing content onto the login screen.
- Fix navbar layout: search box vertical stretch, centered search
  with actions pinned right (responsive, scoped to desktop breakpoint),
  logo/icon overflow that pushed the logo into the breadcrumb bar,
  and the header logo/search overlap.
- Fix the app-switcher flyout menu: BS5's Popper was overriding the
  intended full-height side-panel CSS with popover-style positioning
  (fixed via data-bs-display="static"), hover-to-open wasn't wired to
  BS5's dropdown API, and duplicate carets (manual icon + BS5's auto
  arrow) showed on both the app-switcher and Quick Create > Document.

Systemic JS bugs (root cause of "everything is broken"):
- jQuery's .on('shown.bs.modal', ...) style bindings never fire under
  Bootstrap 5, since BS5 dispatches events using the full dotted
  string as the literal event type while jQuery strips it as a
  namespace. Converted ~25 call sites across helper.js, Vtiger.js,
  List.js, DashBoard.js, Calendar.js, AdvanceSearchList.js,
  SearchList.js, Settings/Vtiger/Index.js, MassEdit.js, EmailPreview.js,
  Inventory/Edit.js, and Products/Detail.js to native addEventListener.
- BS5's Modal reads element.querySelector('.modal-dialog') once at
  construction and calls querySelector on it unconditionally later;
  vtiger's pre-Bootstrap overlay containers (#overlayPage,
  #overlayPageContent, #helpPageOverlay, .myModal, #popupModal) never
  had that wrapper, so it threw and silently aborted the show sequence
  before 'shown.bs.modal' could ever fire - breaking Advanced Search,
  Task Management, and anything else gated on that event. Fixed by
  ensuring an invisible decoy .modal-dialog exists before construction.
- Close buttons across ~20 templates were invisible/tiny (BS5 dropped
  the .close class outright) and non-functional on #overlayPage/
  #helpPageOverlay (missing .modal class broke Bootstrap's
  closest('.modal') dismiss-button lookup); fixed both, and disabled
  the resulting backdrop on those two since they're slide-down panels,
  not centered dialogs, and a backdrop there just intercepts clicks.
- Legacy custom dropdown-toggle handlers (stopImmediatePropagation +
  a homegrown 'open' class) were silently blocking BS5's own delegated
  click handler on Quick Create > Document; removed in favor of native
  BS5 dropdown behavior plus hover-to-open.
- Fixed a fixed-width/table-layout race in the listview's floatThead
  plugin that could zero out every data column but the first.

Environment/setup:
- Regenerate composer.lock for PHP 8.4/8.5 compatibility.
- Fix two PHP 8.1+ deprecations (implicit nullable parameter, and two
  (boolean)/(double) casts modernized to (bool)/(float)).

+++ include/Webservices/LineItem/VtigerLineItemOperation.php
@@ -502,7 +502,7 @@ class VtigerLineItemOperation extends VtigerActorOperation {
 			}
 		}
 
-		if(!empty($parent['hdnDiscountAmount']) && ((double)$parent['hdnDiscountAmount']) > 0){
+		if(!empty($parent['hdnDiscountAmount']) && ((float)$parent['hdnDiscountAmount']) > 0){
 			$discount = ($parent['hdnDiscountAmount']);
 		} elseif(!empty($parent['hdnDiscountPercent'])){
 			$discount = ($parent['hdnDiscountPercent']/100 * $parent['hdnSubTotal']);
